Skip database creation when we are running a production workflow.
Normally, the workflow would be connected to the 'reflow_manager' database so the current_database() would return 'reflow_manager' in a dev environment. If we are in a production environment, reflow would connect to the target database created by the DBA so the name would match the dbName param from the workflow.